namespace reorganisation for middlewares
