+++ Vulpix/src/Vulpix/Startup.cs
@@ -1,11 +1,12 @@
 using Vulpix;
- public class Startup : VulpixCore
-  {
 
-    public Startup(){
-      var config = VulpixConfiguration.GetConfiguration();
-      base.SetRoute(config.GetRoute());
-      base.SetMiddleware(config.GetMiddleware());
-      base.SetPublicFolder(config.GetPublicFolder());
+public class Startup : VulpixCore
+{
+    public Startup()
+    {
+        var config = VulpixConfiguration.GetConfiguration();
+        base.SetRoute(config.GetRoute());
+        base.SetMiddleware(config.GetMiddleware());
+        base.SetPublicFolder(config.GetPublicFolder());
     }
-  }
+}
